requirements.txt
click==8.1.7
colorama==0.4.6
Flask==2.2.5
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyodbc==4.0.39
python-dotenv==0.21.1
typing_extensions==4.7.1
Werkzeug==2.2.3
wincertstore==0.2
zipp==3.15.0
